requirements.txt
requests[socks]>=2.28,<2.29
redis>=4.3,<4.4
Flask>=2.2,<2.3
gunicorn>=20.1,<20.2
python-dotenv>=0.21,<0.22
cloudflare>=2.10,<2.11
PyJWT>=2.6,<2.7
